requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.5
cryptography==45.0.7
daphne==4.2.1
Django==5.2.5
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.1
flower==2.0.1
gunicorn==23.0.0
humanize==4.13.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kafka-python==2.2.15
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
djangorestframework_simplejwt==5.5.1